refactor: use affix component for prefix and suffix components
