+++ src/components/organisms/UiList/_internal/UiListItemPrefix.vue
@@ -1,27 +1,31 @@
 <template>
-  <component
-    :is="listItemPrefixComponent"
-    :class="[
-      'ui-list-item-prefix',
-      { 'ui-button--text': isButton },
-    ]"
+  <UiListItemAffix
+    class="ui-list-item-prefix"
   >
-    <!-- @slot Use this slot to replace Prefix template. -->
-    <slot>
-      <!-- @slot Use this slot to replace icon template. -->
+    <template
+      #default="{
+        icon,
+        label,
+        isButton,
+        hasIcon,
+        iconAttrs,
+        labelAttrs,
+      }"
+    >
       <slot
         name="icon"
         v-bind="{
           hasIcon,
-          iconAttrs: defaultProps.iconAttrs,
+          iconAttrs,
         }"
       >
         <UiIcon
           v-if="hasIcon"
-          v-bind="defaultProps.iconAttrs"
+          :icon="icon"
+          v-bind="iconAttrs"
+          :class="[ { 'ui-button__icon': isButton } ]"
         />
       </slot>
-      <!-- @slot Use this slot to replace label template. -->
       <slot
         name="label"
         v-bind="{
@@ -30,85 +34,22 @@
         }"
       >
         <span
-          v-if="label"
           v-bind="labelAttrs"
-        >
-          {{ label }}
-        </span>
+        >{{ label }}</span>
       </slot>
-    </slot>
-  </component>
+    </template>
+  </UiListItemAffix>
 </template>
 
 <script setup lang="ts">
-import {
-  computed,
-  defineAsyncComponent,
-  useAttrs,
-} from 'vue';
-import type {
-  DefineAttrsProps,
-  HTMLTag,
-  Icon,
-} from '@/types';
-import UiIcon, { type IconAttrsProps } from '../../../atoms/UiIcon/UiIcon.vue';
+import UiListItemAffix, { type ListItemAffixAttrsProps } from '@/components/organisms/UiList/_internal/UiListItemAffix.vue';
+import UiIcon from '../../../atoms/UiIcon/UiIcon.vue';
+import type { DefineAttrsProps } from '../../../../types';
 
-export interface ListItemPrefixProps {
-  /**
-   * Use this props to set list item prefix tag.
-   */
-  tag?: HTMLTag;
-  /**
-   * Use this props to set label.
-   */
-  label?: string;
-  /**
-   * Use this props to set icon.
-   */
-  icon?: Icon;
-  /**
-   * Use this props to pass attrs for prefix UiIcon.
-   */
-  iconAttrs?: IconAttrsProps;
-  /**
-   * Use this props to pass attrs for label element.
-   */
-  labelAttrs?: DefineAttrsProps<null>;
-}
-export type ListItemPrefixAttrsProps = DefineAttrsProps<ListItemPrefixProps>;
-const props = withDefaults(defineProps<ListItemPrefixProps>(), {
-  tag: undefined,
-  label: '',
-  icon: '',
-  iconAttrs: () => ({}),
-  labelAttrs: () => ({}),
-});
-const attrs:ListItemPrefixAttrsProps = useAttrs();
-const isButton = computed(() => (!!Object.keys(attrs)
-  .find((key) => key
-    .match(/(^on*|to|href)/))));
-const listItemPrefixComponent = computed(() => {
-  if (props.tag) {
-    return props.tag;
-  }
-  return isButton.value
-    ? defineAsyncComponent(() => import('../../../atoms/UiButton/UiButton.vue'))
-    : defineAsyncComponent(() => import('../../../atoms/UiText/UiText.vue'));
-});
-const defaultProps = computed(() => ({
-  iconAttrs: {
-    icon: props.icon,
-    class: [
-      { 'ui-button__icon': isButton.value },
-      props.iconAttrs?.class,
-    ],
-    ...props.iconAttrs,
-  },
-}));
-const hasIcon = computed(() => (defaultProps.value?.iconAttrs?.icon));
+export type ListItemPrefixAttrsProps = DefineAttrsProps<ListItemAffixAttrsProps>
 </script>
 
-<style lang="scss">
+<style scoped lang="scss">
 @use "../../../../styles/functions";
 @use "../../../../styles/mixins";
 
@@ -116,57 +57,12 @@ const hasIcon = computed(() => (defaultProps.value?.iconAttrs?.icon));
   $this: &;
   $element: list-item-prefix;
 
-  display: flex;
-  align-items: center;
-  gap: functions.var($element, gap, var(--space-4));
-
-  &:not(.ui-button) {
-    --icon-color: #{functions.var($element, icon-color, var(--color-icon-primary))};
-    --text-color: #{functions.var($element, text-color, var(--color-text-action-primary))};
-  }
-
-  @at-root .ui-list-item__content {
-    @include mixins.hover {
-      #{$this}:not(.ui-button) {
-        --icon-color: #{functions.var($element + '-hover', icon-color, var(--color-icon-primary-hover))};
-        --text-color: #{functions.var($element + '-hover', text-color, var(--color-text-action-primary-hover))};
-      }
-    }
-
-    &:active {
-      #{$this}:not(.ui-button) {
-        --icon-color: #{functions.var($element + '-active', icon-color, var(--color-icon-primary-active))};
-        --text-color: #{functions.var($element, '-active', var(--color-text-action-primary-active))};
-      }
-    }
-  }
-
-  @at-root [class*="-secondary"] {
-    #{$this},
-    &#{$this} {
-      &:not(.ui-button) {
-        --icon-color: #{functions.var($element, icon-color, var(--color-icon-secondary))};
-        --text-color: #{functions.var($element, text-color, var(--color-text-action-secondary))};
-      }
-    }
-
-    @at-root .ui-list-item__content {
-      @include mixins.hover {
-        [class*="-secondary"]#{$this}:not(.ui-button),
-        & [class*="-secondary"] #{$this}:not(.ui-button) {
-          --icon-color: #{functions.var($element + '-hover', icon-color, var(--color-icon-secondary-hover))};
-          --text-color: #{functions.var($element + '-hover', text-color, var(--color-text-action-secondary-hover))};
-        }
-      }
-
-      &:active {
-        [class*="-secondary"]#{$this}:not(.ui-button),
-        & [class*="-secondary"] #{$this}:not(.ui-button) {
-          --icon-color: #{functions.var($element + '-hover', icon-color, var(--color-icon-secondary-active))};
-          --text-color: #{functions.var($element + '-hover', text-color, var(--color-text-action-secondary-active))};
-        }
-      }
-    }
-  }
+  --list-item-affix-gap: functions.var($element, gap);
+  --list-item-affix-icon-color: #{functions.var($element, icon-color)};
+  --list-item-affix-hover-icon-color: #{functions.var($element + '-hover', icon-color)};
+  --list-item-affix-active-icon-color: #{functions.var($element + '-active', icon-color)};
+  --list-item-affix-text-color: #{functions.var($element, text-color)};
+  --list-item-affix-hover-text-color: #{functions.var($element + '-hover', text-color)};
+  --list-item-affix-active-text-color: #{functions.var($element + '-active', text-color)};
 }
 </style>
