+++ src/components/organisms/UiList/_internal/UiListItemSuffix.vue
@@ -1,82 +1,52 @@
 <template>
-  <component
-    :is="listItemSuffixComponent"
-    :class="[
-      'ui-list-item-suffix',
-      { 'ui-button--text': isButton },
-    ]"
+  <UiListItemAffix
+    class="ui-list-item-suffix"
   >
-    <!-- @slot Use this slot to replace suffix template. -->
-    <slot>
-      <!-- @slot Use this slot to replace label template. -->
+    <template
+      #default="{
+        icon,
+        label,
+        isButton,
+        hasIcon,
+        iconAttrs,
+        labelAttrs,
+      }"
+    >
       <slot
         name="label"
         v-bind="{
           label,
-          labelSuffixAttrs,
           labelAttrs,
         }"
       >
         <span
-          v-if="label"
-          v-bind="labelSuffixAttrs || labelAttrs"
-        >
-          {{ label }}
-        </span>
+          v-bind="labelAttrs"
+        >{{ label }}</span>
       </slot>
-      <!-- @slot Use this slot to replace icon template. -->
       <slot
         name="icon"
         v-bind="{
           hasIcon,
-          iconSuffixAttrs: iconSuffixAttrs,
-          iconAttrs: defaultProps.iconAttrs,
+          iconAttrs,
         }"
       >
         <UiIcon
           v-if="hasIcon"
-          v-bind="defaultProps.iconAttrs"
+          :icon="icon"
+          v-bind="iconAttrs"
+          :class="[ { 'ui-button__icon ui-button__icon--end': isButton } ]"
         />
       </slot>
-    </slot>
-  </component>
+    </template>
+  </UiListItemAffix>
 </template>
 
 <script setup lang="ts">
-import {
-  computed,
-  defineAsyncComponent,
-  useAttrs,
-} from 'vue';
+import UiListItemAffix, { type ListItemAffixAttrsProps } from '@/components/organisms/UiList/_internal/UiListItemAffix.vue';
 import UiIcon, { type IconAttrsProps } from '../../../atoms/UiIcon/UiIcon.vue';
-
-import type {
-  DefineAttrsProps,
-  HTMLTag,
-  Icon,
-} from '../../../../types';
+import type { DefineAttrsProps } from '../../../../types';
 
 export interface ListItemSuffixProps {
-  /**
-   * Use this props to set list item suffix tag.
-   */
-  tag?: HTMLTag;
-  /**
-   * Use this props to set label.
-   */
-  label?: string;
-  /**
-   * Use this props to set icon.
-   */
-  icon?: Icon;
-  /**
-   * Use this props to pass attrs for suffix UiIcon.
-   */
-  iconAttrs?: IconAttrsProps;
-  /**
-   * Use this props to pass attrs for label element.
-   */
-  labelAttrs?: DefineAttrsProps<null>;
   /**
    * @deprecated will be removed in 2.0.0 Use this props to pass attrs for suffix UiIcon.
    */
@@ -86,20 +56,13 @@ export interface ListItemSuffixProps {
    */
   labelSuffixAttrs?: DefineAttrsProps<null>;
 }
-export type ListItemSuffixAttrsProps = DefineAttrsProps<ListItemSuffixProps>;
-
-const props = withDefaults(defineProps<ListItemSuffixProps>(), {
-  tag: undefined,
-  label: '',
-  icon: '',
-  iconAttrs: () => ({}),
-  labelAttrs: () => ({}),
-  // @deprecated will be removed in 2.0.0 / BEGIN
+export type ListItemSuffixAttrsProps = DefineAttrsProps<ListItemAffixAttrsProps, ListItemSuffixProps>
+const props = withDefaults(defineProps<ListItemSuffixProps>(), { // @deprecated will be removed in 2.0.0 / BEGIN
   iconSuffixAttrs: () => ({}),
   labelSuffixAttrs: () => ({}),
-  // END
+// END
 });
-const attrs:ListItemSuffixAttrsProps = useAttrs();
+
 // TODO: will be removed in 2.0.0 / BEGIN
 if (Object.keys(props.iconSuffixAttrs).length > 0) {
   console.warn('[@infermedica/component-library]: The `iconSuffixAttrs` props is deprecated and it will be removed in v2.0.0. Please use `iconAttrs` instead.');
@@ -108,34 +71,9 @@ if (Object.keys(props.labelSuffixAttrs).length > 0) {
   console.warn('[@infermedica/component-library]: The `labelSuffixAttrs` props is deprecated and it will be removed in v2.0.0. Please use `labelAttrs` instead.');
 }
 // END
-
-const isButton = computed(() => (!!Object.keys(attrs)
-  .find((key) => key
-    .match(/(^on*|to|href)/))));
-const listItemSuffixComponent = computed(() => {
-  if (props.tag) {
-    return props.tag;
-  }
-  return isButton.value
-    ? defineAsyncComponent(() => import('../../../atoms/UiButton/UiButton.vue'))
-    : defineAsyncComponent(() => import('../../../atoms/UiText/UiText.vue'));
-});
-const defaultProps = computed(() => ({
-  iconAttrs: {
-    icon: props.icon,
-    class: [
-      { 'ui-button__icon ui-button__icon-end': isButton.value },
-      props.iconAttrs?.class,
-    ],
-    ...props.iconAttrs,
-    // @TODO: will be removed in 2.0.0
-    ...props.iconSuffixAttrs,
-  },
-}));
-const hasIcon = computed(() => (defaultProps.value?.iconAttrs?.icon));
 </script>
 
-<style lang="scss">
+<style scoped lang="scss">
 @use "../../../../styles/functions";
 @use "../../../../styles/mixins";
 
@@ -143,57 +81,12 @@ const hasIcon = computed(() => (defaultProps.value?.iconAttrs?.icon));
   $this: &;
   $element: list-item-suffix;
 
-  display: flex;
-  align-items: center;
-  gap: functions.var($element, gap, var(--space-4));
-
-  &:not(.ui-button) {
-    --icon-color: #{functions.var($element, icon-color, var(--color-icon-primary))};
-    --text-color: #{functions.var($element, text-color, var(--color-text-action-primary))};
-  }
-
-  @at-root .ui-list-item__content {
-    @include mixins.hover {
-      #{$this}:not(.ui-button) {
-        --icon-color: #{functions.var($element + '-hover', icon-color, var(--color-icon-primary-hover))};
-        --text-color: #{functions.var($element + '-hover', text-color, var(--color-text-action-primary-hover))};
-      }
-    }
-
-    &:active {
-      #{$this}:not(.ui-button) {
-        --icon-color: #{functions.var($element + '-active', icon-color, var(--color-icon-primary-active))};
-        --text-color: #{functions.var($element, '-active', var(--color-text-action-primary-active))};
-      }
-    }
-  }
-
-  @at-root [class*="-secondary"] {
-    #{$this},
-    &#{$this} {
-      &:not(.ui-button) {
-        --icon-color: #{functions.var($element, icon-color, var(--color-icon-secondary))};
-        --text-color: #{functions.var($element, text-color, var(--color-text-action-secondary))};
-      }
-    }
-
-    @at-root .ui-list-item__content {
-      @include mixins.hover {
-        [class*="-secondary"]#{$this}:not(.ui-button),
-        & [class*="-secondary"] #{$this}:not(.ui-button) {
-          --icon-color: #{functions.var($element + '-hover', icon-color, var(--color-icon-secondary-hover))};
-          --text-color: #{functions.var($element + '-hover', text-color, var(--color-text-action-secondary-hover))};
-        }
-      }
-
-      &:active {
-        [class*="-secondary"]#{$this}:not(.ui-button),
-        & [class*="-secondary"] #{$this}:not(.ui-button) {
-          --icon-color: #{functions.var($element + '-hover', icon-color, var(--color-icon-secondary-active))};
-          --text-color: #{functions.var($element + '-hover', text-color, var(--color-text-action-secondary-active))};
-        }
-      }
-    }
-  }
+  --list-item-affix-gap: functions.var($element, gap);
+  --list-item-affix-icon-color: #{functions.var($element, icon-color)};
+  --list-item-affix-hover-icon-color: #{functions.var($element + '-hover', icon-color)};
+  --list-item-affix-active-icon-color: #{functions.var($element + '-active', icon-color)};
+  --list-item-affix-text-color: #{functions.var($element, text-color)};
+  --list-item-affix-hover-text-color: #{functions.var($element + '-hover', text-color)};
+  --list-item-affix-active-text-color: #{functions.var($element + '-active', text-color)};
 }
 </style>
